+++ tests/Tasks/CheckTest.php
@@ -1,6 +1,7 @@
 <?php
 namespace Rocketeer\Tasks;
 
+use Rocketeer\Tasks\Check;
 use Rocketeer\TestCases\RocketeerTestCase;
 
 class CheckTest extends RocketeerTestCase
@@ -9,9 +10,32 @@ class CheckTest extends RocketeerTestCase
 	{
 		$this->assertTaskHistory('Check', array(
 			'git --version',
-			'{php} -r "print PHP_VERSION;"',
 			'{php} -m',
 		));
+	}
+
+	public function testCanCheckPhpVersion()
+	{
+		$check = new Check($this->app);
+
+		$this->mock('files', 'Filesystem', function ($mock) {
+			return $mock
+				->shouldReceive('put')
+				->shouldReceive('glob')->andReturn(array())
+				->shouldReceive('exists')->andReturn(true)
+				->shouldReceive('get')->andReturn('{"require":{"php":">=5.3.0"}}');
+		});
+		$this->assertTrue($check->checkPhpVersion());
+
+		// This is is going to come bite me in the ass in 10 years
+		$this->mock('files', 'Filesystem', function ($mock) {
+			return $mock
+				->shouldReceive('put')
+				->shouldReceive('glob')->andReturn(array())
+				->shouldReceive('exists')->andReturn(true)
+				->shouldReceive('get')->andReturn('{"require":{"php":">=5.9.0"}}');
+		});
+		$this->assertFalse($check->checkPhpVersion());
 	}
 
 	public function testCanCheckPhpExtensions()
@@ -24,7 +48,6 @@ class CheckTest extends RocketeerTestCase
 
 		$this->assertTaskHistory('Check', array(
 			'git --version',
-			'{php} -r "print PHP_VERSION;"',
 			'{php} -m',
 		));
 	}